refactor(welcome): move plan cards into section-pricing, drop selection
Move plan-selector/plan-option out of the signup modal into
section-pricing/ (renamed section-pricing.vue -> index.vue) so the
welcome pricing section can reuse them directly instead of a bespoke
grid. Since the pricing section only displays plans, strip the
selection paradigm entirely: plan-option is now an inert card with a
name/price header and a feature list, themed via data-theme instead of
a theme prop.
